requirements.txt
tensorflow
pandas
scikit-learn
matplotlib
Flask
Pillow
h5py
